@@ -1,13 +1,12 @@
 using System.Threading.Tasks;
 
-namespace MyCompany.ProjectNameHere.MyDocumentApi.Services {
+namespace MyCompany.ProjectNameHere.MyDocumentApi.Services; 
 
-    public class FieldTemplatingHelper {
+public class FieldTemplatingHelper {
 
-        public async Task<string> ApplySuffixAsync(int input, string suffix) {
+    public async Task<string> ApplySuffixAsync(int input, string suffix) {
 
-            return input + "-" + suffix;
-        }
-
+        return input + "-" + suffix;
     }
+
 }
